+++ ft/ft_printf.c
@@ -28,9 +28,9 @@ int	ft_flags(char const *str, va_list args)
 	else if (*str == 'u')
 		count = ft_unsign(va_arg(args, unsigned int), 0);
 	else if (*str == 'x')
-		count = ft_lowerhex(va_arg(args, unsigned long), 0);
+		count = ft_lowerhex(va_arg(args, unsigned int), 0);
 	else if (*str == 'X')
-		count = ft_upperhex(va_arg(args, unsigned long), 0);
+		count = ft_upperhex(va_arg(args, unsigned int), 0);
 	else if (*str == 'p')
 	{
 		write(1, "0x", 2);
